Revert "LPS-68886 Back port"
This reverts commit 9f8463733ab58f139abd7b94989e928471b962e0.
