+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-service/bnd.bnd
@@ -3,7 +3,7 @@ Bundle-SymbolicName: com.liferay.portal.workflow.kaleo.service
 Bundle-Version: 2.0.13
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Kaleo Workflow Engine
-Liferay-Require-SchemaVersion: 1.3.3
+Liferay-Require-SchemaVersion: 1.3.2
 Liferay-Service: true
 -dsannotations-options: inherit
 -includeresource: @jabsorb-[0-9]*.jar
