+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/KaleoServiceUpgrade.java
@@ -23,7 +23,6 @@ import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0.UpgradeClassNam
 import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0.UpgradeKaleoAction;
 import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0.UpgradeKaleoDefinition;
 import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_2.UpgradeKaleoClassNameAndKaleoClassPK;
-import com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_3.UpgradeBlogsEntryClassName;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -67,10 +66,6 @@ public class KaleoServiceUpgrade implements UpgradeStepRegistrator {
 		registry.register(
 			"com.liferay.portal.workflow.kaleo.service", "1.3.1", "1.3.2",
 			new UpgradeKaleoClassNameAndKaleoClassPK());
-
-		registry.register(
-			"com.liferay.portal.workflow.kaleo.service", "1.3.2", "1.3.3",
-			new UpgradeBlogsEntryClassName());
 	}
 
 }
