+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-service/src/main/java/com/liferay/portal/workflow/kaleo/internal/upgrade/v1_3_0/UpgradeClassNames.java
@@ -14,20 +14,43 @@
 
 package com.liferay.portal.workflow.kaleo.internal.upgrade.v1_3_0;
 
+import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.LoggingTimer;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.upgrade.util.Table;
 import com.liferay.portal.workflow.kaleo.runtime.util.WorkflowContextUtil;
 
 import java.io.Serializable;
+
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 import java.util.Map;
 
 /**
  * @author Lino Alves
  */
-public class UpgradeClassNames extends BaseUpgradeClassNames {
+public class UpgradeClassNames extends UpgradeProcess {
 
 	@Override
+	protected void doUpgrade() throws Exception {
+		updateClassName("KaleoInstance", "className");
+		updateClassName("KaleoInstanceToken", "className");
+		updateClassName("KaleoLog", "currentAssigneeClassName");
+		updateClassName("KaleoLog", "previousAssigneeClassName");
+		updateClassName("KaleoNotificationRecipient", "recipientClassName");
+		updateClassName("KaleoTaskAssignment", "assigneeClassName");
+		updateClassName("KaleoTaskAssignmentInstance", "assigneeClassName");
+		updateClassName("KaleoTaskInstanceToken", "className");
+
+		updateWorkflowContextEntryClassName("KaleoInstance", "kaleoInstanceId");
+		updateWorkflowContextEntryClassName("KaleoLog", "kaleoLogId");
+		updateWorkflowContextEntryClassName(
+			"KaleoTaskInstanceToken", "kaleoTaskInstanceTokenId");
+		updateWorkflowContextEntryClassName(
+			"KaleoTimerInstanceToken", "kaleoTimerInstanceTokenId");
+	}
+
 	protected void updateClassName(String tableName, String columnName) {
 		try (LoggingTimer loggingTimer = new LoggingTimer(tableName)) {
 			Table table = new Table(tableName);
@@ -42,26 +65,63 @@ public class UpgradeClassNames extends BaseUpgradeClassNames {
 		}
 	}
 
-	@Override
-	protected Map<String, Serializable> updateWorkflowContext(
-		String workflowContextJSON) {
+	protected void updateWorkflowContext(
+			String tableName, String primaryKeyName, long primaryKeyValue,
+			String workflowContext)
+		throws Exception {
 
-		String updatedWorkflowContextJSON =
-			_workflowContextUpgradeHelper.renamePortalClassNames(
-				workflowContextJSON);
+		try (PreparedStatement ps = connection.prepareStatement(
+				"update " + tableName + " set workflowContext = ? where " +
+					primaryKeyName + " = ?")) {
 
-		Map<String, Serializable> workflowContext = WorkflowContextUtil.convert(
-			updatedWorkflowContextJSON);
+			ps.setString(1, workflowContext);
+			ps.setLong(2, primaryKeyValue);
 
-		if (workflowContextJSON.equals(updatedWorkflowContextJSON) &&
-			!_workflowContextUpgradeHelper.isEntryClassNameRenamed(
-				workflowContext)) {
-
-			return null;
+			ps.executeUpdate();
 		}
+	}
 
-		return _workflowContextUpgradeHelper.renameEntryClassName(
-			workflowContext);
+	protected void updateWorkflowContextEntryClassName(
+			String tableName, String primaryKeyName)
+		throws Exception {
+
+		try (LoggingTimer loggingTimer = new LoggingTimer(tableName);
+			PreparedStatement ps = connection.prepareStatement(
+				"select " + primaryKeyName + ", workflowContext from " +
+					tableName + " where workflowContext is not null");
+			ResultSet rs = ps.executeQuery()) {
+
+			while (rs.next()) {
+				long primaryKeyValue = rs.getLong(primaryKeyName);
+				String workflowContextJSON = rs.getString("workflowContext");
+
+				if (Validator.isNull(workflowContextJSON)) {
+					continue;
+				}
+
+				String updatedWorkflowContextJSON =
+					_workflowContextUpgradeHelper.renamePortalClassNames(
+						workflowContextJSON);
+
+				Map<String, Serializable> workflowContext =
+					WorkflowContextUtil.convert(updatedWorkflowContextJSON);
+
+				if (workflowContextJSON.equals(updatedWorkflowContextJSON) &&
+					!_workflowContextUpgradeHelper.isEntryClassNameRenamed(
+						workflowContext)) {
+
+					continue;
+				}
+
+				workflowContext =
+					_workflowContextUpgradeHelper.renameEntryClassName(
+						workflowContext);
+
+				updateWorkflowContext(
+					tableName, primaryKeyName, primaryKeyValue,
+					WorkflowContextUtil.convert(workflowContext));
+			}
+		}
 	}
 
 	private final WorkflowContextUpgradeHelper _workflowContextUpgradeHelper =
